Adjusting action when deleting category or author with their books available

+++ controllers/authors.js
@@ -35,6 +35,21 @@ exports.delete = async (req, res, next) => {
   const {
     params: { authorId },
   } = req;
+
+  const authorBook = Book.findOne({ author: authorId });
+  const [bookErr, bookData] = await asyncWrapper(authorBook);
+  if (bookErr) {
+    if (!bookErr.statusCode) {
+      bookErr.statusCode = 500;
+    }
+    return next(bookErr);
+  }
+  if (bookData) {
+    console.log(bookData);
+    const error = new Error('This author has some books and cannot be deleted.');
+    error.status = 409;
+    return next(error);
+  }
   const author = Author.findByIdAndDelete(authorId);
   const [authorErr, authorData] = await asyncWrapper(author);
   if (authorErr) {
@@ -133,24 +148,26 @@ exports.getById = async (req, res, next) => {
   res.status(200).json({ message: 'Author found successfully!', author: authorData });
 };
 
-
 exports.getAuthorBooks = async (req, res, next) => {
   const page = req.query.page || 1;
   const perPage = req.query.perPage || 6;
-  const {authorId}=req.params;
-  let total = await Book.find({author:authorId}).count();
-  let books = Book.find({author:authorId}).populate({
-    path: 'author',
-    select: 'firstName lastName -_id'
-  }).skip((page - 1) * perPage)
-  .limit(perPage);
-  const [booksErr,BooksData] = await asyncWrapper(books);
-  if(booksErr) {
-    if(!booksErr.statusCode){
-      booksErr.statusCode=500;
+  const { authorId } = req.params;
+  let total = await Book.find({ author: authorId }).count();
+  let books = Book.find({ author: authorId })
+    .populate({
+      path: 'author',
+      select: 'firstName lastName -_id',
+    })
+    .skip((page - 1) * perPage)
+    .limit(perPage);
+  const [booksErr, BooksData] = await asyncWrapper(books);
+  if (booksErr) {
+    if (!booksErr.statusCode) {
+      booksErr.statusCode = 500;
     }
-    return next(booksErr)
+    return next(booksErr);
   }
-  res.status(200).json({message:"successfully found Books",authorBooks:BooksData,totalBooks:total});
-}
-
+  res
+    .status(200)
+    .json({ message: 'successfully found Books', authorBooks: BooksData, totalBooks: total });
+};
